Further decoupling of MVC and miscellaneous refactoring
This commit touches almost all parts of the codebase, but the primary focus was
to decouple flow logic from the view components. This commit has been subject to
multiple merge-conflicts and rebasing issues, therefore it is very wide in scope
and has many rough edges. It also likely breaks serveral parts of the
application.